Create 7. Reverse Integer.cpp
check overflow bcz the int max size is 2,147,483,647 and int min size is -2,147,483,648. If the multiplication or addition result goes beyond this then it will cause overflow error.
Positive Overflow Check:

Before multiplying by 10: Check if reversedNumber is greater than INT_MAX / 10. If true, the next multiplication will overflow.
Before adding the digit: If reversedNumber equals INT_MAX / 10, check if the digit to be added is greater than 7 (since the maximum last digit for a 32-bit integer is 7).
Negative Overflow Check:

Before multiplying by 10: Check if reversedNumber is less than INT_MIN / 10. If true, the next multiplication will overflow.
Before adding the digit: If reversedNumber equals INT_MIN / 10, check if the digit to be added is less than -8 (since the minimum last digit for a 32-bit integer is -8).